+++ packages/web3-providers/src/Sentry/index.ts
@@ -10,8 +10,7 @@ import {
     isDeviceOnWhitelist,
 } from '@masknet/shared-base'
 import { removeSensitiveTelemetryInfo } from '@masknet/web3-shared-base'
-import { isNewerThan } from '../helpers/isNewerThan.js'
-import { isSameVersion } from '../helpers/isSameVersion.js'
+import { isNewerThan, isSameVersion } from '../entry-helpers.js'
 import { TelemetryAPI } from '../entry-types.js'
 
 const IGNORE_ERRORS = [
