+++ packages/web3-providers/src/entry-helpers.ts
@@ -9,6 +9,8 @@ export * from './helpers/getAssetFullName.js'
 export * from './helpers/getNativeAssets.js'
 export * from './helpers/getPaymentToken.js'
 export * from './helpers/parseJSON.js'
+export * from './helpers/isNewerThan.js'
+export * from './helpers/isSameVersion.js'
 export * from './helpers/resolveActivityType.js'
 export * from './helpers/resolveActivityTypeBackgroundColor.js'
 export * from './helpers/captureAsyncTransaction.js'
